Split web preview into two separate commands, add title to default browser
Using repeated keystrokes didn't work correctly due to the current NVDA implementation.
Therefore, we are going to use two shortcuts.
Text that is not HTML is now wrapped into minimal tags.
This ensures that the title 'preview of MarkDown or Html' is shown in the external browser,
instead of 'File:///temp path'.